requirements.txt
flask
flask_cors
pytesseract
opencv-python
pillow
speechrecognition
moviepy
pydub
transformers
sentence-transformers
numpy==1.26.4
openai==0.28
python-dotenv
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0 
pandas
PyPDF2
pandas==2.0.2
nltk
rouge_score
pydub==0.25.1
opencv-contrib-python 
Pillow
